Put all of the classes in the same file.
