+++ examples/models.py
@@ -1,6 +1,60 @@
 #!/usr/bin/env python
-from dataset import Dataset
 
+import json
+
+class Dataset():
+    """ Class used to interact with and existing dataset on the Log Insight server. """
+
+    def __init__(self, id, name, description, constraints, type):
+        self.id = id
+        self.name = name
+        self.description = description
+        self.constraints = constraints
+        self.type = type
+
+    def __repr__(self):
+        """A human-readable representation of the dataset, in constructor form."""
+        return '{cls}(id={x.id!r}, name={x.name!r}, description={x.description!r}, constraints={x.constraints!r}, type={x.type!r}'.format(cls=self.__class__.__name__, x=self)
+
+
+class DatasetSpec():
+    """ Class that defines a dataset that needs to be created. """
+    # TODO: This needs to support constraints.  We may want to add constraints
+    # to datasets as a separate operation.
+   
+    def __init__(self, name, description="", constraints=[]):
+        self.name = name
+        self.description = description
+        self.constraints = constraints
+
+    def json(self):
+        return json.dumps({
+            'name': self.name,
+            'description': self.description,
+            'constraints': self.constraints
+            })
+
+    def __repr__(self):
+        """A human-readable representation of the DatasetSpec, in constructor form."""
+        return '{cls}(name={x.name!r}, description={x.description!r})'.format(cls=self.__class__.__name__, x=self)
+
+ 
+class Capability():
+    """ Class used to interact with and existing capability on the Log Insight server. """
+
+    def __init__(self, id):
+        self.id = id
+
+    def json(self):
+        return json.dumps({
+            'id': self.id
+            })
+
+    def __repr__(self):
+        """A human-readable representation of the capability, in constructor form."""
+        return '{cls}(id={x.id!r})'.format(cls=self.__class__.__name__, x=self)
+
+  
 class Group():
     """ Class that helps interact an existing group on the Log Insight server. """
 
@@ -25,6 +79,14 @@ class GroupSpec():
         self.capabilities = capabilities
         self.description = description
         self.datasets = datasets
+
+    def json(self):
+        return json.dumps({
+            'name': self.name,
+            'description': self.description,
+            'capabilities': self.capabilities,
+            'datasets': self.datasets
+            })
 
     def __repr__(self):
         """A human-readable representation of the group specification, in constructor form."""
